% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_beers.R
\name{calc_beers}
\alias{calc_beers}
\title{Perform Beers Interpolation or Subdivision}
\usage{
calc_beers(values, method = "ordinary")
}
\arguments{
\item{values}{A numeric vector of data points for interpolation or subdivision.}

\item{method}{A character string specifying the method to be used. (default = "ordinary")
Must be one of "ordinary", "modified", "subidvision_ordinary", or "subidvision_modified".
- "ordinary": In interpolation, include original data points unchanged; in subdivision,
  each set of 5 subdivided values sums to the original data point.
- "modified": In interpolation, includes some smoothing with only the first and last points unchanged;
  in subdivision, similar smoothing occurs.}
}
\value{
A numeric vector with interpolated or subdivided values.
}
\description{
This function implements the Beers interpolation or subdivision methods, either in their ordinary or modified forms. It generates interpolated or subdivided points from a given data set. The Beers interpolation method, first introduced in "The Record of the American Institute of Actuaries" (Vol. 34, Part I, 1945), is a six-term formula designed to minimize the fifth differences in interpolated results. More details can be found in "The Methods and Materials of Demography, Volume 2" (page 877).
}
\examples{
calc_beers(c(1, 2, 3, 4, 5), "ordinary")
calc_beers(c(1, 2, 3, 4, 5), "modified")
calc_beers(c(1, 2, 3, 4, 5), "subidvision_ordinary")
calc_beers(c(1, 2, 3, 4, 5), "subidvision_modified")
}
