% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_beers.R
\name{interpolate_beers}
\alias{interpolate_beers}
\title{Fill Gaps Using Beers Method Interpolation}
\usage{
interpolate_beers(input_df, groups, year, values, method = "ordinary")
}
\arguments{
\item{input_df}{A data frame containing the data to be interpolated.}

\item{groups}{A vector of column names to group by.}

\item{year}{The name of the column containing year information.}

\item{values}{The name of the column containing values for interpolation.}

\item{method}{A character string specifying the method to be used.
Must be one of "ordinary", "modified", "subidvision_ordinary", or "subidvision_modified".
- "ordinary" (default): In interpolation, include original data points unchanged; in subdivision,
  each set of 5 subdivided values sums to the original data point.
- "modified": In interpolation, includes some smoothing with only the first and last points unchanged;
  in subdivision, similar smoothing occurs.}
}
\value{
A data frame with interpolated values using the Beers method.
}
\description{
This function applies the Beers method interpolation on a grouped data frame.
It requires each group to have at least 6 non-NA values and assumes data to be in 5-year intervals.
Groups with fewer than 6 non-NA values are excluded from interpolation.
}
\examples{
data <- data.frame(
  Scenario = rep(c("Scenario1", "Scenario2"), each = 6),
  Region = rep(c("Region1", "Region2"), each = 6),
  year = rep(c(2000, 2005, 2010, 2015, 2020, 2025), 2),
  value = rnorm(12)
)

data

filled_data <- gtaptools::interpolate_beers(
  input_df = data,
  groups = c("Scenario", "Region"),
  year = "year",
  values = "value"
)

filled_data

}
