% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_spline.R
\name{interpolate_spline}
\alias{interpolate_spline}
\title{Fill Gaps with Spline Interpolation}
\usage{
interpolate_spline(input_df, groups, year, values, method = "fmm")
}
\arguments{
\item{input_df}{A data frame containing the data to be processed.}

\item{groups}{A vector of column names to group by (to loop applying spline).}

\item{year}{The name of the column containing years.}

\item{values}{The name of the numeric column containing values for interpolation.}

\item{method}{The method of interpolation. Possible values:
   Must be one of "fmm", "natural", "periodic", "monoH.FC" or "hyman".
- "fmm" (default): Forsythe, Malcolm, and Moler method. An exact cubic spline is fitted through
  the four points at each end of the data. This is used for determining end conditions.
  Not suitable for extrapolation.
- "natural": Natural spline method. Uses natural splines for interpolation. Linear
  extrapolation outside the range of x using the slope of the interpolating curve at the
  nearest data point.
- "periodic": Periodic spline method. Suitable for periodic data.
- "monoH.FC": Monotone Hermite spline according to Fritsch and Carlson. Ensures the spline
  is monotone (increasing or decreasing) if the data are monotone.
- "hyman": Monotone cubic spline using Hyman filtering of an "fmm" fit for strictly
  monotonic inputs.}
}
\value{
A data frame with gaps in years filled using spline interpolation.
}
\description{
This function takes a data frame and performs cubic spline interpolation
to fill in missing year gaps for specified groups.
Groups with fewer than 2 non-NA values are excluded from interpolation.
}
\examples{
data <- data.frame(
  Scenario = rep(c("Scenario1", "Scenario2"), each = 5),
  Region = rep(c("Region1", "Region2"), each = 5),
  year = rep(c(2000, 2005, 2010, 2015, 2020), 2),
  value = rnorm(10)
)

data

filled_data <- gtaptools::interpolate_spline(
  input_df = data,
  groups = c("Scenario", "Region"),
  year = "year",
  values = "value"
)

filled_data

}
