% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bars}
\alias{plot_bars}
\title{Bar graphs}
\usage{
plot_bars(
  input_data,
  x,
  x_label = NULL,
  y,
  y_label = NULL,
  fill = NULL,
  facet = NULL,
  facet_n_row = NULL,
  facet_scales = NULL,
  palette = 1,
  orientation_bars = "vertical",
  rotate_x_labels = 90,
  legend_title = NULL,
  legend_pos = "bottom",
  reactive = T,
  gtap_theme = "A"
)
}
\arguments{
\item{input_data}{It must consist of one or more input databases, which must be separated from each other by sublists (see example). In the case of multiple databases, all will be combined for the final output.Arrays and data.frames must be inside sublists (list(....)) as indicated in the examples section. Aggregations on input data can only be performed on single array and data.frame inputs.}

\item{x}{the name of the variable to be plotted on the x-axis.}

\item{x_label}{the label to be used for the x-axis (default is x).}

\item{y}{The name of the variable to be plotted on the y-axis.}

\item{y_label}{the label to be used for the x-axis (default is y).}

\item{fill}{the name of the variable to be used for stacking the bars, if any (default is NULL)}

\item{facet}{the name of the variable to be used for faceting the graph, if any (default is NULL)}

\item{palette}{the name of the color palette to be used. Can be specified by name or number from 1 to about 15 (default is 1).}

\item{orientation_bars}{the orientation of the bars to be plotted, "h" to horizontal or "v" to vertical (default is "v").}

\item{rotate_x_labels}{the angle in degrees to rotate the x-axis labels (default is 90)}

\item{legend_title}{the title to be used for the legend (default is fill).}

\item{legend_pos}{the position to be used for the legend ("bottom", "left" or "right"), default is "bottom".}

\item{reactive}{Plot reactive (default = T).}

\item{gtap_theme}{the gtap template theme. Can be specified from 1 to about 5 (default is 1).}
}
\description{
Plot static and reactive bar charts.
}
