% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot distributions}
\usage{
plot_dist(
  input_data,
  value_var,
  type = "density",
  x_label = "",
  y_label = "",
  fill = NULL,
  facet = NULL,
  linewidth = 0.5,
  palette = "Pastel1",
  transparency = 0.7,
  orientation_bars = "vertical",
  rotate_x_labels = 0,
  legend_title = NULL,
  legend_pos = "right",
  adjust = 2,
  reactive = T,
  gtap_theme = "A"
)
}
\arguments{
\item{input_data}{A data.frame or array object containing the data to be plotted}

\item{value_var}{Numeric variable to plot the distribution.}

\item{type}{Type of distribution chart plot such as "density", "histogram", "violin" or "boxplot" (Default = "density").}

\item{x_label}{the label to be used for the x-axis.}

\item{y_label}{the label to be used for the y-axis.}

\item{fill}{the name of the variable to be used for stacking the distributions, if any (default is NULL)}

\item{facet}{the name of the variable to be used for faceting the graph, if any (default is NULL)}

\item{linewidth}{Line width/thickness (default = 0.5).}

\item{palette}{the name of the color palette to be used. Can be specified by name or number from 1 to about 15 (default is 1).}

\item{transparency}{degree of color transparency [0-1] (default = 0.7).}

\item{orientation_bars}{the orientation of the bars to be plotted, "h" to horizontal or "v" to vertical (default is "v").}

\item{rotate_x_labels}{the angle in degrees to rotate the x-axis labels (default is 90)}

\item{legend_title}{the title to be used for the legend (default is fill).}

\item{legend_pos}{the position to be used for the legend ("bottom", "left" or "right"), default is "bottom".}

\item{adjust}{A multiplicate bandwidth adjustment (default = 2).}

\item{reactive}{Plot reactive (default = TRUE).}

\item{gtap_theme}{the gtap template theme. Can be specified from 1 to about 5 (default is 1).}
}
\description{
Tool to plot distribution charts.
}
