% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_grid.R
\name{plot_map_grid}
\alias{plot_map_grid}
\title{Plot gridded maps}
\usage{
plot_map_grid(
  input_data,
  value_var,
  latitude,
  longitude,
  crs = 4326,
  resolution = NULL,
  interpolate = F,
  colors,
  borders_color = "black",
  borders_size = 1,
  legend_labels = NULL,
  legend_title = NULL,
  legend_pos = NULL,
  reactive = T,
  fillOpacity = 0.8,
  gtap_theme = "A"
)
}
\arguments{
\item{input_data}{An input data.frame that has geographical coordinates information and at least one numeric column}

\item{value_var}{The name of the numeric variable to be plotted on the map.}

\item{latitude}{The name of the numeric variable that contains the latitude information of each grid.}

\item{longitude}{The name of the numeric variable that contains the longitude information of each grid.}

\item{crs}{Coordinate reference system to be assigned to *latitude* and *longitude* (default = 4326, the World Geodetic System 1984 - EPSG:4326).}

\item{resolution}{Resolution of the grids. (default = NULL, automatically estimated based on the distance between coordinates.).}

\item{interpolate}{A model from gstat::gstat (gstat package) to interpolate values using a fitted model object. If TRUE it applies a simple kriging using the formula z~1. (default = F).}

\item{colors}{The name of the color palette to be adopted (see a list in the section Notes below) or the custom color break vector. See the examples section.}

\item{borders_color}{Border line color.}

\item{borders_size}{Line border size.}

\item{legend_labels}{Legend labels. It must be the same length as the color vector.}

\item{legend_title}{Legend title.}

\item{legend_pos}{The position of legends. For reactive = F:"none", "left", "right", "bottom", "top", or two-element numeric vector. For reactive = T: "topright", "bottomright", "bottomleft" and "topleft".}

\item{reactive}{Plot reactive (default = T).}

\item{fillOpacity}{Color fill layer transparency. Must be between [0, 1] (default = 0.8). Only applied when reactive = T.}

\item{gtap_theme}{The gtap template theme. Can be specified from 1 to about 5 (default is 1). Only applied when reactive = F.}
}
\description{
Plot static and reactive spatial grid viz.
}
\note{
Palette options include:
Viridis palletes:
  "magma", "inferno", "plasma", "viridis", "cividis", "rocket", "mako", "turbo".
Color Brewer Sequential:
  "Blues", "BuGn", "BuPu", "GnBu", "Greens", "Greys", "Oranges", "OrRd", "PuBu", "PuBuGn", "PuRd", "Purples", "RdPu", "Reds", "YlGn", "YlGnBu", "YlOrBr", "YlOrRd".
Color Brewer Diverging:
  "BrBG", "PiYG", "PRGn", "PuOr", "RdBu", "RdGy", "RdYlBu", "RdYlGn", "Spectral".
}
