% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\title{Squeeze the simulation folder.}
\usage{
squeeze(cmf_file, zip_file, add_files, output = F, bat = T)
}
\arguments{
\item{cmf_file}{Path to .cmf file which manages the simulation.}

\item{zip_file}{Name of the .zip file that will be created.}

\item{add_files}{Vector with the names or extensions of the files that will also be included in the .zip file in addition to the files mentioned in the description.}

\item{output}{Includes simulation output files (default = F).}

\item{bat}{Create a batch file to compile (if necessary) and run the simulation. (default = V). It is necessary to have Gempack installed.}
}
\description{
Squeeze the simulation folder by selecting only the essential files and putting them in a .zip file. It also creates a .bat file that makes it easy to run the simulation later.The files that are included are those specified in the .cmf file and that have the extension .tab, .cmf, .sti, .bat, .har, .prm, .shk, .cls, and in the case output = F, .sl4, .upd, .slc.
}
