% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze_sim}
\alias{squeeze_sim}
\title{Squeeze the simulation files into a .zip file.}
\usage{
squeeze_sim(cmf_file, zip_file, add_files = NULL, output = F, bat = T)
}
\arguments{
\item{cmf_file}{Path to .cmf file which manages the simulation.}

\item{zip_file}{Name of the .zip file that will be created.}

\item{add_files}{Vector with the names or extensions of the files that will also be included in the .zip file in addition to the files mentioned in the description.}

\item{output}{Includes simulation output files (default = F).}

\item{bat}{Create a batch file to compile (if necessary) and run the simulation (default = T). For this functionality it is necessary to have Gempack installed.}
}
\description{
Scans the .cmf file and selects just essential files for the simulation and compresses them in a .zip file. It also creates a .bat file that makes it easy to run the simulation later. The files that are included are those specified in the .cmf file and that have the extension .tab, .cmf, .sti, .bat, .har, .prm, .shk, .cls, and in the case output = F, .sl4, .upd, .slc. (This function does not support dynamic simulations operationalized in the RunDynam software.)
}
