% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_shape.R
\name{har_shape}
\alias{har_shape}
\title{Merge and generate new headers.}
\usage{
har_shape(
  input_data,
  new_calculated_vars = NULL,
  del_headers = NULL,
  output_har_file = NULL
)
}
\arguments{
\item{input_data}{It must consist of one or more input databases, which must be separated from each other by sublists (see example). In the case of multiple databases, all will be combined for the final output.}

\item{new_calculated_vars}{New variables resulting from custom calculations between the headers contained in input_data. Each variable's parameters must be informed (it can be *x, y, z ...*), the function *fun* that represents the calculation to be done, the *new_header_name*, and the *sets* for the output structure. The different headers must have at least one similar set so that it is possible to establish correspondence between them. Please note the example section.}

\item{del_headers}{Vector of characters with the names of headers that must be excluded from the output.}

\item{output_har_file}{Output .har file name.}
}
\description{
Allows the combination of different databases in data.frame or array format. Generate new variables flexibly from custom functions. Calculations can be performed between headers/variables of different dimensions/sets.
}
