% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_shape.R
\name{har_shape}
\alias{har_shape}
\title{Aggregates headers of data in .har structure.}
\usage{
har_shape(
  input_data,
  new_calculated_vars = NULL,
  del_var = NULL,
  output_har_file = NULL
)
}
\arguments{
\item{input_data}{It can indicate a path to a .har file or an existing object in the R environment that has the output structure of the read_har function.}

\item{new_calculated_vars}{XXXX}

\item{del_var}{XXXX}

\item{output_har_file}{Output .har file name.}
}
\description{
It aggregates variables from a .har file on disk or an object with the structure exported by the read_har function. It is possible to adopt customized weights and functions to calculate aggregations. The specification of GTAP models (...) through the *model* parameter is supported so that the respective weight variables are automatically detected according to the model being analyzed.
}
