% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_header.R
\name{summarise_header}
\alias{summarise_header}
\title{Aggregates headers of data in .har structure.}
\usage{
summarise_header(
  input_data,
  header,
  fun = function(x) sum(x, na.rm = T),
  export_sets = T,
  output_har_file = NULL,
  output_csv_file = NULL,
  ...
)
}
\arguments{
\item{input_data}{An array that has the output structure of the read_har function or a data.frame.}

\item{header}{Must be adopted the format header_name[c("its sets")], where the header_name must be the same name of the numeric values column in case of a data.frame input_data, and the "its sets" must be the same name of the categorical columns of that data.frame. Please check the examples section and the package's online manual for more details.}

\item{fun}{Function used for aggregation in case of non-unique values in sets (default = sum).}

\item{export_sets}{If TRUE, the vectors with the set elements are incorporated with the output. If an output .har file is indicated, it will be created and exported to that .har file. If FALSE, they will not be exported.}

\item{output_har_file}{Output .har file name.}

\item{output_csv_file}{Output .csv file name.}

\item{...}{Any additional arguments to be used to write the .csv file through data.table::fwrite, such as separator character (sep = ","), the character for decimal points (dec = "."), etc.}
}
\value{
An array or vector of characters (sets) structured in a compatible way to compose a .har file.
}
\description{
Summarizes a single database to an array compatible format for writing to .har files.
}
\examples{
# example code



}
