% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_header.R
\name{summarise_header}
\alias{summarise_header}
\title{Aggregates headers of data in .har structure.}
\usage{
summarise_header(
  input_data,
  sets,
  values,
  fun = function(x) sum(x, na.rm = T),
  new_header_name,
  export_sets = T,
  output_har_file = NULL
)
}
\arguments{
\item{input_data}{An array that has the output structure of the read_har function or a data.frame.}

\item{sets}{Dimensions/columns names that contain the categorical variables for summarizing the output sets. It must be a maximum of 3 characters each.}

\item{values}{Name of the variable that contains the numerical values.}

\item{fun}{Function used for aggregation in case of non-unique values in sets (default = sum).}

\item{new_header_name}{Name of the new header created.}

\item{export_sets}{If TRUE, the vectors with the set elements are incorporated with the output. If an output .har file is indicated, it will be created and exported to that .har file. If FALSE, they will not be exported.}

\item{output_har_file}{Output .har file name.}
}
\description{
Summarizes a single database to a compatible format for writing to .har files.
}
